kbuild: add read-file macro
Since GNU Make 4.2, $(file ...) supports the read operater '<', which
is useful to read a file without forking any process. No warning is
shown even if the input file is missing.

For older Make versions, it falls back to the cat command.

The added ifeq will break when GNU Make 4.10 or 10.0 is released.
It will take a long time if the current release pace continues.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Reviewed-by: Nicolas Schier <nicolas@fjasle.eu>