ovl: remove privs in fallocate, copyfile/reflink
When modifiying a file privileges (suid/sgid/security xattr) need to be
removed.  This needs to be done with the original creds; doing it with the
mounter's creds won't have the desired effect.

Add explicit removal of privileges to the fallocate, copyfile and reflink
operations.

This is tested by xfstests/generic/(673,683,684,685,686,687).

The tests still fail with this patch applied, but that's because of the
inconsistency in removing sgid without group exectute.  This has no
security implications and will be addressed separately.

Fixes: aab8848cee5e ("ovl: add ovl_fallocate()")
Fixes: 8ede205541ff ("ovl: add reflink/copyfile/dedup support")
Cc: <stable@vger.kernel.org> # v4.19
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>