arm64: efi: Move runtime services asm wrapper out of .text
The EFI runtime services wrapper code checks whether register X18 was
corrupted by firmware code, and prints a diagnostic message to the
kernel log if this was the case, as this violates the EFI spec, and may
interfere with the OS's use of this register as a platform register (as
the AAPCS calls it). This functionality was added in commit 7e611e7dbb23
("efi/arm64: Check whether x18 is preserved by runtime services calls"),
at which time X18 was not used in such a special role, so we did not
bother to restore it, and only moaned about it.

Only when shadow call stack support was added, we started using X18 in a
special role, to carry the shadow call stack pointer. This meant that,
when X18 does get corrupted, we need to restore it from a known good
value before proceeding, as C code built with shadow call stacks enabled
cannot function without one. And given that we already preserved the old
value in order to determine whether EFI corrupted it or not, we could
restore it just as easily, as we did in commit e73f02c6eb15 ("arm64:
efi: Restore register x18 if it was corrupted")

However, restoring the shadow call stack pointer from the stack is a bad
idea, as the whole point of the shadow call stack is that the ordinary
stack cannot be trusted with return addresses. And having a snippet of
code in the executable mapping of the kernel text that does this is
asking for problems.

So let's move this code out of there, and only map it executable in the
EFI page tables, which are only ever active while a runtime service call
is in progress, and only on the CPU that makes the call.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>