mmc: sdhci-of-at91: add soc data field quirks2
Add a quirks2 field in the soc_data struct, this replaces
the hs200_broken bool. The purpose is to be able to add multiple
quirks in the soc_data without having bools for each of them.
Storing them in a single variable makes the code cleaner.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
(cherry picked from linux-6.1-trunk/at91/mmc)
Signed-off-by: Andrei Simion <andrei.simion@microchip.com>