dts: lan966x-ev23x71a: Fix microchip,sd-sgpio
This property is used to define which external SGPIO port and bit will
detect the signal detect of the SFP port. The current values are wrong,
the value that needs to be written has the format port << 2 | bit. As
per schematic the SGPIO ports are 6,7,8 and 9 and for all the bit is 0.