arm64: dts: microchip: set number of DMA channels
Currently, XDMAC driver is reading the secure register but that seems to
either be empty or set to 1 for the number of DMA channels which is not
correct as controller is supposed to have 16 channels according to the
datasheet.

Now SPI and UART can actually get channels as previously XDMAC only
registered a single DMA channel.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>