btrfs: validate system chunk array at btrfs_validate_super()
Currently btrfs_validate_super() only does a very basic check on the
array chunk size (not too large than the available space, but not too
small to contain no chunk).

The more comprehensive checks (the regular chunk checks and size check
inside the system chunk array) is all done inside
btrfs_read_sys_array().

It's not a big deal, but for the sake of concentrated verification, we
should validate the system chunk array at the time of super block
validation.

So this patch does the following modification:

- Introduce a helper btrfs_check_system_chunk_array()
  * Validate the disk key
  * Validate the size before we access the full chunk/stripe items.
  * Do the full chunk item validation

- Call btrfs_check_system_chunk_array() at btrfs_validate_super()

- Simplify the checks inside btrfs_read_sys_array()
  Now the checks will be converted to an ASSERT().

- Simplify the checks inside read_one_chunk()
  Now all chunk items inside system chunk array and chunk tree is
  verified, there is no need to verify it again inside read_one_chunk().

This change has the following advantages:

- More comprehensive checks at write time
  Although this also means extra memcpy() for the superblocks at write
  time, due to the limits that we need a dummy extent buffer to utilize
  all the extent buffer helpers.

- Slightly improved readablity when iterating the system chunk array

Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>