nfs: use wait_var_event_spinlock() to wait for nfsi->layout to change.
The recently added interface wait_var_event_spinlock() is designed for
waiting for events that can only be checked under a spinlock.
This matches the requirements for waiting for nfsi->layout so we can
change to using wait_var_event_spinlock().

This avoids a use of plain wake_up_var() which always needs to be
checked for correct barrier usage.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Anna Schumaker <anna.schumaker@oracle.com>