maple_tree: make MT_FLAGS_LOCK_IRQ do something
I'm not sure what the original intention was but I take it that it wsas
to indicate that the lock to be taken must be irq safe. I need this for
pidfs as it's called from alloc_pid() which expects irq safe locking.
Make mtree_{un}lock() check MT_FLAGS_LOCK_IRQ and if present call
spin_{un}lock_irq().

Link: https://lore.kernel.org/r/20241209-work-pidfs-maple_tree-v2-1-003dbf3bd96b@kernel.org
Signed-off-by: Christian Brauner <brauner@kernel.org>