PCI: dwc: Clean up some unnecessary codes in dw_pcie_suspend_noirq()
Before sending PME_TURN_OFF, don't test the LTSSM state. Since it's
safe to send PME_TURN_OFF message regardless of whether the link is
up or down.

Thus, there is no need to test the LTSSM state before sending the
PME_TURN_OFF message, and only print the message when ltssm_stat is
not in DETECT and POLL. There shouldn't be an error message shown
when no endpoint is connected at all.

Link: https://lore.kernel.org/r/20241210081557.163555-3-hongxing.zhu@nxp.com
Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
[kwilczynski: commit log]
Signed-off-by: Krzysztof Wilczyński <kwilczynski@kernel.org>
Reviewed-by: Frank Li <Frank.Li@nxp.com>
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>