misc: pci_endpoint_test: Fix overflow of bar_size
Running 'pcitest -b 2' fails with "TEST FAILED" when the BAR2 size
exceeds 4GB.

The return value of the pci_resource_len() macro can be larger than that
of a signed integer type. As such, for example, when using 'pcitest'
with an 8GB BAR2, the bar_size of the integer type will overflow.

Thus, change the data type of bar_size from integer to resource_size_t,
to fix the integer overflow for large BAR sizes.

Link: https://lore.kernel.org/r/20250109094556.1724663-3-18255117159@163.com
Signed-off-by: Hans Zhang <18255117159@163.com>
[kwilczynski: commit log]
Signed-off-by: Krzysztof Wilczyński <kwilczynski@kernel.org>
Reviewed-by: Niklas Cassel <cassel@kernel.org>
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>